// https_service_docker_volume/www/modules/Tree/src/TreeMetadata.d.ts
export declare class TreeMetadata {
    NODE_ICON__SRC: string;
    NODE_LABEL__TEXT: string;
    SUBTREE: string;
    NODE__ID: string;
    NODE__HOLDER_ID: string;
    NODE__PATH: string;
    NODE__ORDER: string;
    NODE__HYPERLINK: string;
    NODE__OPENED: string;
    NODE__CSS_CLASS_NAME: string;
    NODE__ART: string;
    constructor();
}